+++ api-derive/council/votes.ts
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.6';
-import type { QueryableModuleStorage } from 'https://deno.land/x/polkadot@0.2.2/api-base/types/index.ts';
-import type { Vec } from 'https://deno.land/x/polkadot@0.2.2/types/mod.ts';
-import type { AccountId, Balance } from 'https://deno.land/x/polkadot@0.2.2/types/interfaces/index.ts';
-import type { PalletElectionsPhragmenVoter } from 'https://deno.land/x/polkadot@0.2.2/types/lookup.ts';
-import type { ITuple } from 'https://deno.land/x/polkadot@0.2.2/types/types/index.ts';
+import type { QueryableModuleStorage } from 'https://deno.land/x/polkadot/api-base/types/index.ts';
+import type { Vec } from 'https://deno.land/x/polkadot/types/mod.ts';
+import type { AccountId, Balance } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { PalletElectionsPhragmenVoter } from 'https://deno.land/x/polkadot/types/lookup.ts';
+import type { ITuple } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { DeriveApi, DeriveCouncilVote, DeriveCouncilVotes } from '../types.ts';
 
 import { combineLatest, map, of } from 'https://esm.sh/rxjs@7.5.6';