+++ api-derive/democracy/util.ts
@@ -1,12 +1,12 @@
 // Copyright 2017-2022 @polkadot/api-derive authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { ReferendumInfoTo239, Tally } from 'https://deno.land/x/polkadot@0.2.2/types/interfaces/index.ts';
-import type { PalletDemocracyReferendumInfo, PalletDemocracyReferendumStatus, PalletDemocracyVoteThreshold } from 'https://deno.land/x/polkadot@0.2.2/types/lookup.ts';
-import type { Option } from 'https://deno.land/x/polkadot@0.2.2/types-codec/mod.ts';
+import type { ReferendumInfoTo239, Tally } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { PalletDemocracyReferendumInfo, PalletDemocracyReferendumStatus, PalletDemocracyVoteThreshold } from 'https://deno.land/x/polkadot/types/lookup.ts';
+import type { Option } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
 import type { DeriveReferendum, DeriveReferendumVote, DeriveReferendumVotes, DeriveReferendumVoteState } from '../types.ts';
 
-import { BN, bnSqrt, objectSpread } from 'https://deno.land/x/polkadot@0.2.2/util/mod.ts';
+import { BN, bnSqrt, objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 interface ApproxState {
   votedAye: BN;