+++ rpc-core/util/drr.ts
@@ -5,7 +5,7 @@ import type { Observable } from 'https://esm.sh/rxjs@7.5.6';
 
 import { catchError, distinctUntilChanged, publishReplay, refCount, tap } from 'https://esm.sh/rxjs@7.5.6';
 
-import { logger, stringify } from 'https://deno.land/x/polkadot@0.2.2/util/mod.ts';
+import { stringify } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { refCountDelay } from './refCountDelay.ts';
 
@@ -17,18 +17,17 @@ interface Options {
   skipTimeout?: boolean;
 }
 
-const l = logger('drr');
+function CMP (a: unknown, b: unknown): boolean {
+  return stringify({ t: a }) === stringify({ t: b });
+}
 
-const CMP = (a: unknown, b: unknown): boolean =>
-  stringify({ t: a }) === stringify({ t: b });
-
-const ERR = (error: Error): Observable<never> => {
-  l.error(error.message);
-
+function ERR (error: Error): Observable<never> {
   throw error;
-};
+}
 
-const NOOP = (): void => undefined;
+function NOOP (): void {
+  // empty
+}
 
 /**
  * Shorthand for distinctUntilChanged(), publishReplay(1) and refCount().
@@ -36,14 +35,16 @@ const NOOP = (): void => undefined;
  * @ignore
  * @internal
  */
-export const drr = ({ delay, skipChange = false, skipTimeout = false }: Options = {}): DrrResult => <T> (source$: Observable<T>): Observable<T> =>
-  source$.pipe(
-    catchError(ERR),
-    skipChange
-      ? tap(NOOP)
-      : distinctUntilChanged<T>(CMP),
-    publishReplay(1),
-    skipTimeout
-      ? refCount()
-      : refCountDelay(delay)
-  );
+export function drr ({ delay, skipChange = false, skipTimeout = false }: Options = {}): DrrResult {
+  return <T> (source$: Observable<T>): Observable<T> =>
+    source$.pipe(
+      catchError(ERR),
+      skipChange
+        ? tap(NOOP)
+        : distinctUntilChanged<T>(CMP),
+      publishReplay(1),
+      skipTimeout
+        ? refCount()
+        : refCountDelay(delay)
+    );
+}