+++ types/interfaces/support/types.ts
@@ -1,8 +1,8 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { Struct, bool, u8 } from 'https://deno.land/x/polkadot@0.2.2/types-codec/mod.ts';
-import type { Balance, Perbill } from 'https://deno.land/x/polkadot@0.2.2/types/interfaces/runtime/index.ts';
+import type { Struct, bool, u8 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { Balance, Perbill } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
 
 /** @name WeightToFeeCoefficient */
 export interface WeightToFeeCoefficient extends Struct {